added the limelight cam turning on every time the limelight is switch to and when "shooter mode" is enabled.

+++ frc1388-2020/src/main/java/frc/robot/subsystems/ShooterSubsystem.java
@@ -308,7 +308,7 @@ public class ShooterSubsystem extends SubsystemBase {
 
       // log RPM
       if( DriverStation.getInstance().isAutonomous() ){
-        USBLogging.info("Target = " + m_rpm + "||RPM = " + getShooterRPM());
+        USBLogging.info("Target = " + m_rpm + "||RPM = " + (int) getShooterRPM());
       }
     }else{
       m_shootMotor.set(0);
